refactor(core): separate large test suites
